Mark zeroconf.utils as protected by renaming to zeroconf._utils (#582)
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
