+++ zeroconf/_dns.py
@@ -27,6 +27,9 @@ from typing import Any, Dict, List, Optional, TYPE_CHECKING, Tuple, Union, cast
 
 from ._exceptions import AbstractMethodException, IncomingDecodeError, NamePartTooLongException
 from ._logger import QuietLogger, log
+from ._utils.net import _is_v6_address
+from ._utils.struct import int2byte
+from ._utils.time import current_time_millis, millis_to_seconds
 from .const import (
     _CLASSES,
     _CLASS_MASK,
@@ -49,9 +52,6 @@ from .const import (
     _TYPE_SRV,
     _TYPE_TXT,
 )
-from .utils.net import _is_v6_address
-from .utils.struct import int2byte
-from .utils.time import current_time_millis, millis_to_seconds
 
 
 if TYPE_CHECKING:
